backend/requirements.txt
Django==5.2.8
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
django-filter==23.5
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
nltk==3.8.1
spacy==3.7.2
pyyaml==6.0.1
requests==2.31.0
celery==5.3.6
redis==5.0.1